feat: Add npm publishing workflow
This commit introduces a new GitHub Actions workflow for publishing
the package to npm. It also includes a new script for building the npm
package using the Deno dnt tool. The version of the package is passed
as an argument to the script. The package.json for the npm package is
generated by the dnt tool during the build process.

The .gitignore file has been updated to ignore the npm directory which
is created during the build process.

The import_map.json file has been updated to include the Deno dnt tool
as a dependency.

The deno.json file has been updated to include a new task for building
the npm package.