feat: update dependencies and test framework
This commit updates the dependencies in deno.jsonc and deno.lock files.
It also replaces the Deno test framework with @cross/test in all test
files. The assert module is updated to @std/assert. These changes are
necessary to keep the project up-to-date with the latest libraries and
best practices.